Fold BaseActivity into MainActivity
closes #404
